add some error handling

+++ src/controllers/moods.ts
@@ -1,17 +1,25 @@
 import { Request, Response, NextFunction } from "express";
+import { validationResult } from "express-validator";
 import moodsService from "../services/moods";
 
 // TODO: req.session.userId is not defined but it should be as path is protected
 // https://stackoverflow.com/questions/66614337/typescript-req-user-is-possibly-undefined-express-and-passport-js
 
 const getAll = async (req: Request, res: Response, next: NextFunction) => {
-  console.log("getting all moods");
   const { userId } = req.session;
-  try {
-    const moods = await moodsService.getAll({ userId });
-    res.render("index", { moods, title: "Moods", user: userId });
 
-    // res.json(moods);
+  try {
+    const moods: any = await moodsService.getAll({ userId });
+    // parse the mood date to a readable format
+    moods.forEach((mood: any) => {
+      mood.created_at = new Date(mood.created_at).toLocaleDateString("en-GB", {
+        year: "numeric",
+        month: "long",
+        day: "numeric",
+      });
+    });
+
+    res.render("index", { moods, title: "Moods", user: userId });
   } catch (e) {
     console.error("Error getting all moods: ", e);
     next(e);
@@ -22,16 +30,29 @@ const getOne = async (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params;
   const { userId } = req.session;
 
-  console.log("Do i run??");
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    res.status(400).json({ errors: errors.array() });
+    return;
+  }
 
   try {
-    const mood = await moodsService.getOne({ id, userId });
-    if (!mood.length) {
+    const [mood]: any = await moodsService.getOne({ id, userId });
+
+    if (!mood) {
       return res.status(404).json({ message: "Mood not found" });
     }
-    res.render("mood", { mood: mood[0], title: "Mood", user: userId });
+
+    mood.created_at = new Date(mood.created_at).toLocaleDateString("en-GB", {
+      year: "numeric",
+      month: "long",
+      day: "numeric",
+    });
+
+    res.render("mood", { mood, title: "Mood", user: userId });
   } catch (e) {
     console.error("Error getting one mood: ", e);
+    res.status(404).json({ message: "Mood not found" + e });
     next(e);
   }
 };
@@ -45,11 +66,28 @@ const create = async (req: Request, res: Response, next: NextFunction) => {
   const { userId } = req.session;
   const { value, context } = req.body;
 
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    res.status(400).json({ errors: errors.array() });
+    return;
+  }
+
   try {
-    const mood = await moodsService.create({ value, context, userId });
-    res.redirect("/moods");
-  } catch (e) {
+    const moodId = await moodsService.create({ value, context, userId });
+    res.status(201).redirect("/moods");
+  } catch (e: any) {
     console.error("Error creating mood: ", e);
+    const customErrors = [];
+
+    switch (e.message) {
+      case "Could not create mood":
+        customErrors.push({ msg: "Could not create mood" });
+        break;
+      default:
+        customErrors.push({ msg: "Error creating a mood" + e });
+    }
+
+    res.status(400).json({ errors: errors.array() });
     next(e);
   }
 };
@@ -59,13 +97,28 @@ const update = async (req: Request, res: Response, next: NextFunction) => {
   const { userId } = req.session;
   const { context } = req.body;
 
-  console.log("UPDATING XDDDD");
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    res.status(400).json({ errors: errors.array() });
+    return;
+  }
 
   try {
     const moodId = await moodsService.update({ id, context, userId });
-    res.json(moodId);
-  } catch (e) {
+    res.json(moodId).status(200);
+  } catch (e: any) {
     console.error("Error updating mood: ", e);
+    const customErrors = [];
+
+    switch (e.message) {
+      case "Could not update mood":
+        customErrors.push({ msg: "Could not update mood" });
+        break;
+      default:
+        customErrors.push({ msg: "Error updating a mood" + e });
+    }
+
+    res.status(400).json({ errors: customErrors });
     next(e);
   }
 };
@@ -74,13 +127,28 @@ const remove = async (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params;
   const { userId } = req.session;
 
-  console.log("REMOVINGG XDDDD");
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    res.status(400).json({ errors: errors.array() });
+    return;
+  }
 
   try {
     const moodId = await moodsService.remove({ id, userId });
     res.json(moodId);
-  } catch (e) {
+  } catch (e: any) {
     console.error("Error removing mood: ", e);
+    const customErrors = [];
+
+    switch (e.message) {
+      case "Could not remove mood":
+        customErrors.push({ msg: "Could not remove mood" });
+        break;
+      default:
+        customErrors.push({ msg: "Error removing a mood" + e });
+    }
+
+    res.status(400).json({ errors: customErrors });
     next(e);
   }
 };
